* subversion/tests/svn_test_main.c   (main): Cast away an integer width mismatch warning.
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1325238 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/tests/svn_test_main.c
@@ -461,7 +461,7 @@ main(int argc, const char *argv[])
         case server_minor_version_opt:
           {
             char *end;
-            opts.server_minor_version = strtol(opt_arg, &end, 10);
+            opts.server_minor_version = (int) strtol(opt_arg, &end, 10);
             if (end == opt_arg || *end != '\0')
               {
                 fprintf(stderr, "FAIL: Non-numeric minor version given\n");
